Move _configure_runtime_env back to the CLI
