+++ python/ray/serve/tests/test_cli.py
@@ -12,7 +12,7 @@ from ray import serve
 from ray.tests.conftest import tmp_working_dir  # noqa: F401, E501
 from ray._private.test_utils import wait_for_condition
 from ray.dashboard.optional_utils import RAY_INTERNAL_DASHBOARD_NAMESPACE
-from ray.serve.scripts import process_args_and_kwargs
+from ray.serve.scripts import process_args_and_kwargs, _configure_runtime_env
 
 
 def ping_endpoint(endpoint: str, params: str = ""):
@@ -147,6 +147,72 @@ class TestProcessArgsAndKwargs:
             args, kwargs = process_args_and_kwargs(empty_val)
             assert args == []
             assert kwargs == {}
+
+
+class TestConfigureRuntimeEnv:
+    @serve.deployment
+    def f():
+        pass
+
+    @pytest.mark.parametrize("ray_actor_options", [None, {}])
+    def test_empty_ray_actor_options(self, ray_actor_options):
+        runtime_env = {
+            "working_dir": "http://test.com",
+            "pip": ["requests", "pendulum==2.1.2"],
+        }
+        deployment = TestConfigureRuntimeEnv.f.options(
+            ray_actor_options=ray_actor_options
+        )
+        _configure_runtime_env(deployment, runtime_env)
+        assert deployment.ray_actor_options["runtime_env"] == runtime_env
+
+    def test_overwrite_all_options(self):
+        old_runtime_env = {
+            "working_dir": "http://test.com",
+            "pip": ["requests", "pendulum==2.1.2"],
+        }
+        new_runtime_env = {
+            "working_dir": "http://new.com",
+            "pip": [],
+            "env_vars": {"test_var": "test"},
+        }
+        deployment = TestConfigureRuntimeEnv.f.options(
+            ray_actor_options={"runtime_env": old_runtime_env}
+        )
+        _configure_runtime_env(deployment, new_runtime_env)
+        assert deployment.ray_actor_options["runtime_env"] == new_runtime_env
+
+    def test_overwrite_some_options(self):
+        old_runtime_env = {
+            "working_dir": "http://new.com",
+            "pip": [],
+            "env_vars": {"test_var": "test"},
+        }
+        new_runtime_env = {
+            "working_dir": "http://test.com",
+            "pip": ["requests", "pendulum==2.1.2"],
+        }
+        merged_env = {
+            "working_dir": "http://test.com",
+            "pip": ["requests", "pendulum==2.1.2"],
+            "env_vars": {"test_var": "test"},
+        }
+        deployment = TestConfigureRuntimeEnv.f.options(
+            ray_actor_options={"runtime_env": old_runtime_env}
+        )
+        _configure_runtime_env(deployment, new_runtime_env)
+        assert deployment.ray_actor_options["runtime_env"] == merged_env
+
+    def test_overwrite_no_options(self):
+        runtime_env = {
+            "working_dir": "http://test.com",
+            "pip": ["requests", "pendulum==2.1.2"],
+        }
+        deployment = TestConfigureRuntimeEnv.f.options(
+            ray_actor_options={"runtime_env": runtime_env}
+        )
+        _configure_runtime_env(deployment, {})
+        assert deployment.ray_actor_options["runtime_env"] == runtime_env
 
 
 def test_start_shutdown(ray_start_stop):
